Refactor: Moved lib and utils from helpers to core
