+++ .luarc.json
@@ -3,7 +3,6 @@
     "Lua.diagnostics.globals": [
         "_neorgcmd_generate_completions",
         "_neorg_module_autocommand_triggered",
-        "neorg",
-        "vim",
+        "vim"
     ]
 }
