+++ lua/neorg/modules/core/esupports/hop/module.lua
@@ -12,8 +12,9 @@ prompted with a set of actions that you can perform on the broken link.
 --]]
 
 local neorg = require("neorg.core")
+local lib, utils = neorg.lib, neorg.utils
+
 require("neorg.modules.base") -- TODO: Move to its own local core module
-require("neorg.external.helpers") -- TODO: Move to its own local core module
 
 local module = neorg.modules.create("core.esupports.hop")
 
@@ -119,7 +120,7 @@ module.public = {
         end
 
         if located_link_information then
-            neorg.lib.match(located_link_information.type)({
+            lib.match(located_link_information.type)({
                 -- If we're dealing with a URI, simply open the URI in the user's preferred method
                 external_app = function()
                     os_open_link(located_link_information.uri)
@@ -337,7 +338,7 @@ module.public = {
             return
         end
 
-        local query = neorg.utils.ts_parse_query("norg", query_str)
+        local query = utils.ts_parse_query("norg", query_str)
 
         for id, node in query:iter_captures(document_root, 0) do
             local capture = query.captures[id]
@@ -428,7 +429,7 @@ module.public = {
             return
         end
 
-        local query = neorg.utils.ts_parse_query("norg", query_text)
+        local query = utils.ts_parse_query("norg", query_text)
         local range = module.required["core.integrations.treesitter"].get_node_range(link_node)
 
         local parsed_link_information = {
@@ -449,12 +450,12 @@ module.public = {
                 and capture_node_range.column_end <= range.column_end
             then
                 local extract_node_text =
-                    neorg.lib.wrap(module.required["core.integrations.treesitter"].get_node_text, node)
+                    lib.wrap(module.required["core.integrations.treesitter"].get_node_text, node)
 
                 parsed_link_information[capture] = parsed_link_information[capture]
-                    or neorg.lib.match(capture)({
+                    or lib.match(capture)({
                         link_file_text = extract_node_text,
-                        link_type = neorg.lib.wrap(string.sub, node:type(), string.len("link_target_") + 1),
+                        link_type = lib.wrap(string.sub, node:type(), string.len("link_target_") + 1),
                         link_location_text = extract_node_text,
                         link_description = extract_node_text,
 
@@ -496,7 +497,7 @@ module.public = {
             end
         end
 
-        return neorg.lib.match(parsed_link_information.link_type)({
+        return lib.match(parsed_link_information.link_type)({
             url = function()
                 return { type = "external_app", uri = parsed_link_information.link_location_text }
             end,
@@ -512,7 +513,7 @@ module.public = {
                     vim.tbl_contains({ "/", "~" }, destination:sub(1, 1)) and "" or (vim.fn.expand("%:p:h") .. "/")
                 ) .. destination
 
-                return neorg.lib.match(vim.fn.fnamemodify(destination, ":e"))({
+                return lib.match(vim.fn.fnamemodify(destination, ":e"))({
                     [{ "jpg", "jpeg", "png", "pdf" }] = {
                         type = "external_app",
                         uri = vim.uri_from_fname(vim.fn.expand(destination)),
@@ -562,7 +563,7 @@ module.public = {
             end,
 
             _ = function()
-                local query_str = neorg.lib.match(parsed_link_information.link_type)({
+                local query_str = lib.match(parsed_link_information.link_type)({
                     generic = [[
                         (_
                           [(strong_carryover_set
@@ -604,7 +605,7 @@ module.public = {
                                 (multi_%s_prefix)
                                   title: (paragraph_segment) @title)])
                         ]],
-                        neorg.lib.reparg(parsed_link_information.link_type, 5)
+                        lib.reparg(parsed_link_information.link_type, 5)
                     ),
                     _ = string.format(
                         [[
@@ -622,7 +623,7 @@ module.public = {
                               (%s_prefix)
                               title: (paragraph_segment) @title)
                         ]],
-                        neorg.lib.reparg(parsed_link_information.link_type, 2)
+                        lib.reparg(parsed_link_information.link_type, 2)
                     ),
                 })
 
@@ -632,7 +633,7 @@ module.public = {
                     return
                 end
 
-                local query = neorg.utils.ts_parse_query("norg", query_str)
+                local query = utils.ts_parse_query("norg", query_str)
 
                 for id, node in query:iter_captures(document_root, buf_pointer) do
                     local capture = query.captures[id]
@@ -745,7 +746,7 @@ module.private = {
     ---@param parsed_link_information table #A table as returned by `parse_link()`
     ---@return table #A table of similarities (fuzzed items)
     fix_link_strict = function(parsed_link_information)
-        local query = neorg.lib.match(parsed_link_information.link_type)({
+        local query = lib.match(parsed_link_information.link_type)({
             generic = [[
                 (_
                   [(strong_carryover_set
@@ -786,7 +787,7 @@ module.private = {
                         (multi_%s_prefix)
                           title: (paragraph_segment) @title)])
             ]],
-                neorg.lib.reparg(parsed_link_information.link_type, 5)
+                lib.reparg(parsed_link_information.link_type, 5)
             ),
             _ = string.format(
                 [[
@@ -804,7 +805,7 @@ module.private = {
                         (%s_prefix)
                         title: (paragraph_segment) @title)
                 ]],
-                neorg.lib.reparg(parsed_link_information.link_type, 2)
+                lib.reparg(parsed_link_information.link_type, 2)
             ),
         })
 
@@ -828,7 +829,7 @@ module.private = {
             end
         end
 
-        local query = neorg.utils.ts_parse_query("norg", query_str)
+        local query = utils.ts_parse_query("norg", query_str)
 
         local document_root = module.required["core.integrations.treesitter"].get_document_root(buffer)
 
@@ -855,7 +856,7 @@ module.private = {
         end
 
         if vim.tbl_isempty(similarities) then
-            neorg.utils.notify("Sorry, Neorg couldn't fix that link :(", vim.log.levels.WARN)
+            utils.notify("Sorry, Neorg couldn't fix that link :(", vim.log.levels.WARN)
         end
 
         table.sort(similarities, function(lhs, rhs)
@@ -879,10 +880,10 @@ module.private = {
 
         local range = module.required["core.integrations.treesitter"].get_node_range(link_node)
 
-        local prefix = neorg.lib.when(
+        local prefix = lib.when(
             parsed_link_information.link_type == "generic" and not force_type,
             "#",
-            neorg.lib.match(most_similar.node:type())({
+            lib.match(most_similar.node:type())({
                 heading1 = "*",
                 heading2 = "**",
                 heading3 = "***",
@@ -910,17 +911,17 @@ module.private = {
 
         callback(
             "{"
-                .. neorg.lib.when(
+                .. lib.when(
                     parsed_link_information.link_file_text,
-                    neorg.lib.lazy_string_concat(":", parsed_link_information.link_file_text, ":"),
+                    lib.lazy_string_concat(":", parsed_link_information.link_file_text, ":"),
                     ""
                 )
                 .. prefix
                 .. most_similar.text
                 .. "}"
-                .. neorg.lib.when(
+                .. lib.when(
                     parsed_link_information.link_description,
-                    neorg.lib.lazy_string_concat("[", parsed_link_information.link_description, "]"),
+                    lib.lazy_string_concat("[", parsed_link_information.link_description, "]"),
                     ""
                 )
         )
