+++ lua/neorg/modules/core/highlights/module.lua
@@ -9,6 +9,8 @@ Neorg under a single tree of highlights: `@neorg.*`.
 --]]
 
 local neorg = require("neorg.core")
+local lib = neorg.lib
+
 require("neorg.modules.base") -- TODO: Move to its own local core module
 
 local module = neorg.modules.create("core.highlights")
@@ -476,7 +478,7 @@ module.public = {
             local is_link = highlight:sub(1, 1) == "+"
 
             local full_highlight_name = "@neorg" .. prefix .. (hl_name:len() > 0 and ("." .. hl_name) or "")
-            local does_hl_exist = neorg.lib.inline_pcall(vim.api.nvim_exec, "highlight " .. full_highlight_name, true)
+            local does_hl_exist = lib.inline_pcall(vim.api.nvim_exec, "highlight " .. full_highlight_name, true)
 
             -- If we are dealing with a link then link the highlights together (excluding the + symbol)
             if is_link then
@@ -513,7 +515,7 @@ module.public = {
             end
 
             local full_highlight_name = "@neorg" .. prefix .. (hl_name:len() > 0 and ("." .. hl_name) or "")
-            local does_hl_exist = neorg.lib.inline_pcall(vim.api.nvim_exec, "highlight " .. full_highlight_name, true)
+            local does_hl_exist = lib.inline_pcall(vim.api.nvim_exec, "highlight " .. full_highlight_name, true)
 
             -- If the highlight already exists then assume the user doesn't want it to be
             -- overwritten
